+++ FalloutStrategy/FalloutEngine/GenericPlayer.cpp
@@ -30,7 +30,7 @@ bool GenericPlayer::DeleteEntity(j1Entity* entity) {
 	switch (entity->type)
 	{
 	case MELEE:
-		for(size_t m = 0; m < troops.size(); m++)
+		for(int m = 0; m < troops.size(); m++)
 		{
 			if (troops[m] == (Troop*)entity) {
 				troops.erase(troops.begin() + m);
@@ -39,7 +39,7 @@ bool GenericPlayer::DeleteEntity(j1Entity* entity) {
 		melees--;
 		break;
 	case RANGED:
-		for(size_t r = 0; r < troops.size(); r++)
+		for(int r = 0; r < troops.size(); r++)
 		{
 			if (troops[r] == (Troop*)entity) {
 				troops.erase(troops.begin() + r);
@@ -48,7 +48,7 @@ bool GenericPlayer::DeleteEntity(j1Entity* entity) {
 		rangeds--;
 		break;
 	case GATHERER:
-		for(size_t g = 0; g < gatherers_vector.size(); g++) {
+		for(int g = 0; g < gatherers_vector.size(); g++) {
 			if (gatherers_vector[g] == (Gatherer*)entity) {
 				gatherers_vector.erase(gatherers_vector.begin() + g);
 			}
