+++ FalloutStrategy/FalloutEngine/j1EntityManager.cpp
@@ -68,7 +68,7 @@ bool j1EntityManager::Awake(pugi::xml_node& config){
 	std::string base_folder = animation_node.attribute("base_folder").as_string();
 	pugi::xml_node file_node = animation_node.child("file");
 
-	for(size_t i = 0; i < REFERENCE_ENTITIES; i++)
+	for(int i = 0; i < REFERENCE_ENTITIES; i++)
 	{
 		texture_folders[i] = base_folder;
 		texture_folders[i].append(file_node.attribute("folder").as_string());
@@ -77,9 +77,9 @@ bool j1EntityManager::Awake(pugi::xml_node& config){
 		reference_entities[i] = nullptr;
 	}
 
-	for(size_t y = 0; y < 150; y++)
+	for(int y = 0; y < 150; y++)
 	{
-		for(size_t x = 0; x < 150; x++)
+		for(int x = 0; x < 150; x++)
 		{
 			occupied_tiles[x][y] = false;
 		}
@@ -102,9 +102,9 @@ bool j1EntityManager::Start() {
 	loading_faction = VAULT;
 	loading_entity = MELEE;
 
-	for(size_t y = 0; y < 150; y++)
+	for(int y = 0; y < 150; y++)
 	{
-		for(size_t x = 0; x < 150; x++)
+		for(int x = 0; x < 150; x++)
 		{
 			occupied_tiles[x][y] = false;
 		}
@@ -112,9 +112,9 @@ bool j1EntityManager::Start() {
 
 	//automatic entities loading
 	int i = 0;
-	for(size_t faction = VAULT; faction < NO_FACTION; faction++)
+	for(int faction = VAULT; faction < NO_FACTION; faction++)
 	{
-		for(size_t type = MELEE; type < BIGHORNER; type++)
+		for(int type = MELEE; type < BIGHORNER; type++)
 		{
 			reference_entities[i] = CreateEntity((Faction)faction, (EntityType)type, faction, type);
 			i++;
@@ -146,7 +146,7 @@ bool j1EntityManager::CleanUp()
 	bool ret = true;
 
 	// -- Instance
-	for(size_t i = 0; i < REFERENCE_ENTITIES; i++)
+	for(int i = 0; i < REFERENCE_ENTITIES; i++)
 	{
 		if (reference_entities[i] != nullptr) {
 			App->tex->UnLoad(reference_entities[i]->texture);
@@ -156,7 +156,7 @@ bool j1EntityManager::CleanUp()
 	}
 
 	// -- Entities
-	for (size_t i = 0; i < entities.size(); i++)
+	for(int i = 0; i < entities.size(); i++)
 	{
 		if (entities[i] != nullptr) {
 			delete entities[i];
@@ -167,7 +167,7 @@ bool j1EntityManager::CleanUp()
 	entities.clear();
 
 	// -- Particles
-	for (size_t i = 0; i < particles.size(); i++)
+	for(int i = 0; i < particles.size(); i++)
 	{
 		if (particles[i] != nullptr) {
 			delete particles[i];
@@ -182,7 +182,7 @@ bool j1EntityManager::CleanUp()
 	smoke = nullptr;
 
 	// -- Buildings
-	for (size_t j = 0; j < resource_buildings.size(); j++)
+	for(int j = 0; j < resource_buildings.size(); j++)
 	{
 		delete resource_buildings[j];
 		resource_buildings[j] = nullptr;
@@ -195,7 +195,7 @@ bool j1EntityManager::CleanUp()
 bool j1EntityManager::PreUpdate() {
 	bool ret = true;
 
-	for (size_t i = 0; i < entities.size(); i++)
+	for(int i = 0; i < entities.size(); i++)
 	{
 		if ((entities[i]->target_entity != nullptr) && (entities[i]->target_entity->to_delete)) {
 			entities[i]->target_entity = nullptr;
@@ -240,7 +240,7 @@ bool j1EntityManager::Update(float dt)
 
 	if (!App->isPaused)
 	{
-		for (size_t i = 0; i < entities.size(); i++)
+		for(int i = 0; i < entities.size(); i++)
 		{
 			if(!entities[i]->to_delete)
 				entities[i]->Update(dt);
@@ -261,9 +261,9 @@ bool j1EntityManager::PostUpdate()
 	{
 		if (App->render->debug) {
 			//resource buildings debug
-			for (size_t i = 0; i < resource_buildings.size(); i++)
+			for(int i = 0; i < resource_buildings.size(); i++)
 			{
-				for (size_t j = 0; j < resource_buildings[i]->tiles.size(); j++)
+				for(int j = 0; j < resource_buildings[i]->tiles.size(); j++)
 				{
 					SDL_Rect rect = { 128,0,64,64 };
 					tex_position = App->map->MapToWorld(resource_buildings[i]->tiles[j].x, resource_buildings[i]->tiles[j].y);
@@ -393,7 +393,7 @@ bool j1EntityManager::PostUpdate()
 			sort_timer.Start();
 		}
 
-		for (size_t i = 0; i < entities.size(); i++)
+		for(int i = 0; i < entities.size(); i++)
 		{
 			//camera culling
 			if ((entities[i]->position.x + entities[i]->sprite_size * 0.5f > -App->render->camera.x)
@@ -408,9 +408,9 @@ bool j1EntityManager::PostUpdate()
 
 		if (App->render->debug) {
 			iPoint occuppied_tile = { -1,-1 };
-			for(size_t y = 0; y < 150; y++)
+			for(int y = 0; y < 150; y++)
 			{
-				for(size_t x = 0; x < 150; x++)
+				for(int x = 0; x < 150; x++)
 				{
 					if (occupied_tiles[x][y]) {
 						occuppied_tile = App->map->MapToWorld(x, y);
@@ -625,14 +625,14 @@ bool j1EntityManager::LoadReferenceEntityData() {
 void j1EntityManager::DestroyEntity(j1Entity* entity) { entity->to_delete = true;}
 
 void j1EntityManager::DestroyAllEntities() {
-	for (size_t i = 0; i < entities.size(); i++)
+	for(int i = 0; i < entities.size(); i++)
 	{
 		entities[i]->to_delete = true;
 	}
 }
 
 void j1EntityManager::DestroyAllEntitiesNow() {
-	for (size_t i = 0; i < entities.size(); i++)
+	for(int i = 0; i < entities.size(); i++)
 	{
 		delete entities[i];
 		entities[i] = nullptr;
@@ -641,7 +641,7 @@ void j1EntityManager::DestroyAllEntitiesNow() {
 }
 
 j1Entity* j1EntityManager::FindEntityByTile(iPoint tile) {
-	for (size_t i = 0; i < entities.size(); i++)
+	for(int i = 0; i < entities.size(); i++)
 	{
 		if (entities[i]->is_dynamic)
 		{
@@ -651,7 +651,7 @@ j1Entity* j1EntityManager::FindEntityByTile(iPoint tile) {
 		else
 		{
 			StaticEntity* static_entity = (StaticEntity*)entities[i];
-			for (size_t j = 0; j < static_entity->tiles.size(); j++)
+			for(int j = 0; j < static_entity->tiles.size(); j++)
 			{
 				if (static_entity->tiles[j] == tile)
 					return entities[i];
@@ -662,9 +662,9 @@ j1Entity* j1EntityManager::FindEntityByTile(iPoint tile) {
 }
 
 ResourceBuilding* j1EntityManager::FindResourceBuildingByTile(iPoint tile) {
-	for (size_t i = 0; i < resource_buildings.size(); i++)
+	for(int i = 0; i < resource_buildings.size(); i++)
 	{
-		for (size_t j = 0; j < resource_buildings[i]->tiles.size(); j++)
+		for(int j = 0; j < resource_buildings[i]->tiles.size(); j++)
 		{
 			if (resource_buildings[i]->tiles[j] == tile)
 				return resource_buildings[i];
@@ -675,7 +675,7 @@ ResourceBuilding* j1EntityManager::FindResourceBuildingByTile(iPoint tile) {
 
 iPoint j1EntityManager::ClosestTile(iPoint position, std::vector<iPoint> entity_tiles) {
 	iPoint pivot = entity_tiles[0];
-	for (size_t i = 0; i < entity_tiles.size(); i++)
+	for(int i = 0; i < entity_tiles.size(); i++)
 	{
 		if (position.DistanceManhattan(entity_tiles[i]) < position.DistanceManhattan(pivot))
 			pivot = entity_tiles[i];
@@ -691,9 +691,9 @@ iPoint j1EntityManager::FindFreeAdjacentTile(iPoint origin, iPoint destination) 
 	int distance_to_destination = 100000;
 
 	while (max < 5) {
-		for(size_t y = -max; y <= max; y++)
+		for(int y = -max; y <= max; y++)
 		{
-			for(size_t x = -max; x <= max; x++)
+			for(int x = -max; x <= max; x++)
 			{
 				if (x != 0 || y != 0) {
 					possible_tile.x = destination.x + x;
@@ -722,7 +722,7 @@ iPoint j1EntityManager::FindFreeAdjacentTile(iPoint origin, iPoint destination) 
 ResourceBuilding* j1EntityManager::GetClosestResourceBuilding(iPoint current_position) {
 	ResourceBuilding* closest_building = nullptr;
 	int min_distance = 1000;
-	for (size_t i = 0; i < resource_buildings.size(); i++)
+	for(int i = 0; i < resource_buildings.size(); i++)
 	{
 		if (resource_buildings[i]->quantity > 0)
 		{
@@ -764,7 +764,7 @@ void j1EntityManager::RandomFactions() {
 	int temp = 0;
 	int randomIndex = 0;
 
-	for(size_t i = 0; i < 4; i++) {
+	for(int i = 0; i < 4; i++) {
 		randomIndex = rand() % 4;
 		temp = randomFaction[i];
 		randomFaction[i] = randomFaction[randomIndex];
@@ -772,13 +772,13 @@ void j1EntityManager::RandomFactions() {
 	}
 
 
-	for(size_t i = 0; i < 4; i++)
+	for(int i = 0; i < 4; i++)
 		LOG("faction %i", randomFaction[i]);
 }
 
 void j1EntityManager::OnCommand(std::vector<std::string> command_parts) {
 	if (command_parts[0] == "destroy_all_entities") {
-		for (size_t i = 0; i < entities.size(); i++)
+		for(int i = 0; i < entities.size(); i++)
 		{
 			if (entities[i]->is_dynamic)
 				entities[i]->to_delete = true;
@@ -792,8 +792,8 @@ void j1EntityManager::LoadUpgradeCosts(pugi::xml_node& config)
 	Faction faction = NO_FACTION;
 	std::string faction_name;
 
-	for(size_t j = 0; j < 3; j++) {
-		for(size_t i = 0; i < 4; i++) {
+	for(int j = 0; j < 3; j++) {
+		for(int i = 0; i < 4; i++) {
 			if (i == 0)faction = VAULT;
 			else if (i == 1)faction = BROTHERHOOD;
 			else if (i == 2)faction = MUTANT;
@@ -899,7 +899,7 @@ ResourceBuilding* j1EntityManager::CreateResourceSpot(int position_x, int positi
 }
 
 void j1EntityManager::DestroyResourceSpot(ResourceBuilding* resource_spot) {
-	for (size_t i = 0; i < resource_buildings.size(); i++)
+	for(int i = 0; i < resource_buildings.size(); i++)
 	{
 		if (resource_spot == resource_buildings[i]) {
 			delete resource_buildings[i];
@@ -1024,9 +1024,9 @@ bool j1EntityManager::Load(pugi::xml_node& data)
 				width = height = 3;
 			}
 
-			for(size_t y = 0; y < height; y++)
+			for(int y = 0; y < height; y++)
 			{
-				for(size_t x = 0; x < width; x++)
+				for(int x = 0; x < width; x++)
 				{
 					tile.x = current_tile.x + x;
 					tile.y = current_tile.y + y;
@@ -1063,7 +1063,7 @@ bool j1EntityManager::Load(pugi::xml_node& data)
 // Save Game State
 bool j1EntityManager::Save(pugi::xml_node& data) const
 {	
-	for (size_t i = 0; i < entities.size(); i++)
+	for(int i = 0; i < entities.size(); i++)
 	{
 		pugi::xml_node entities_pugi = data.append_child("entity");
 		entities_pugi.append_attribute("number") = i;
@@ -1131,8 +1131,8 @@ iPoint j1EntityManager::FindSpawnPoint(int position_x, int position_y) {
 		bool spawnPointFound = false;
 
 		while (occupied_tiles[position_x][position_y]) {
-			for(size_t k = 0; k < 10; k++) {
-				for(size_t i = 0; i <= 5; i++) {
+			for(int k = 0; k < 10; k++) {
+				for(int i = 0; i <= 5; i++) {
 					if (spawnPointFound == false) {
 						if (!occupied_tiles[position_x - i][position_y + k]) {
 							position_x -= i;
@@ -1142,7 +1142,7 @@ iPoint j1EntityManager::FindSpawnPoint(int position_x, int position_y) {
 					}
 				}
 				if (spawnPointFound == false) {
-					for(size_t j = 0; j <= 5; j++) {
+					for(int j = 0; j <= 5; j++) {
 						if (spawnPointFound == false) {
 							if (!occupied_tiles[position_x + k][position_y - j]) {
 								position_y -= j;
@@ -1212,7 +1212,7 @@ void j1EntityManager::SpawnAnimals() {
 	
 	int randomAnimal = 0;
 
-	for(size_t i = 0; i < 15; i++)
+	for(int i = 0; i < 15; i++)
 	{
 		srand((unsigned int)time(NULL)+i);
 		randomAnimal = rand() % 2;
@@ -1246,7 +1246,7 @@ ParticleSystem* j1EntityManager::CreateParticle(fPoint pos) {
 }
 
 void j1EntityManager::DeleteParticles(){
-	for (size_t i = 0; i < particles.size(); i++)
+	for(int i = 0; i < particles.size(); i++)
 	{
 		delete particles[i];
 		particles[i] = nullptr;
@@ -1256,7 +1256,7 @@ void j1EntityManager::DeleteParticles(){
 
 void j1EntityManager::ReleaseParticle(ParticleSystem* particle) {
 
-	/*for(size_t i = 0; i < particles.size(); i++)
+	/*for(int i = 0; i < particles.size(); i++)
 	{
 		Deleting particles from vector
 		if (particles[i] == particle)
@@ -1269,9 +1269,9 @@ void j1EntityManager::ReleaseParticle(ParticleSystem* particle) {
 }
 
 void j1EntityManager::RestartOccupiedTiles() {
-	for(size_t y = 0; y < 150; y++)
+	for(int y = 0; y < 150; y++)
 	{
-		for(size_t x = 0; x < 150; x++)
+		for(int x = 0; x < 150; x++)
 		{
 			occupied_tiles[x][y] = false;
 		}
